Move Markdown actual styling to bridge/standalone packages
Having them all in the same package caused problems when trying to apply
the style in the code; in particular, when your prod code uses bridge
but the tests use standalone, you'd end up with both variants on the
classpath, and it would be impossible to choose which one to use when.

This moves all bridge theming to .bridge, and all standalone theming in
.standalone, to clearly distinguish them. I also renamed the various
theming files to fit the common naming pattern we use elsewhere in the
library.
