+++ markdown/ide-laf-bridge-styling/api/ide-laf-bridge-styling.api
@@ -1,10 +1,13 @@
-public final class org/jetbrains/jewel/intui/markdown/BridgeMarkdownBlockRendererExtensionsKt {
-	public static final fun ProvideMarkdownStyling (Ljava/lang/String;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Lorg/jetbrains/jewel/markdown/processing/MarkdownProcessor;Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
+public final class org/jetbrains/jewel/intui/markdown/bridge/BridgeMarkdownBlockRendererExtensionsKt {
 	public static final fun create (Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Ljava/util/List;Lorg/jetbrains/jewel/markdown/rendering/InlineMarkdownRenderer;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;
 	public static synthetic fun create$default (Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Ljava/util/List;Lorg/jetbrains/jewel/markdown/rendering/InlineMarkdownRenderer;ILjava/lang/Object;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;
 }
 
-public final class org/jetbrains/jewel/intui/markdown/styling/MarkdownBridgeStylingKt {
+public final class org/jetbrains/jewel/intui/markdown/bridge/BridgeProvideMarkdownStylingKt {
+	public static final fun ProvideMarkdownStyling (Ljava/lang/String;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Lorg/jetbrains/jewel/markdown/processing/MarkdownProcessor;Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
+}
+
+public final class org/jetbrains/jewel/intui/markdown/bridge/styling/BridgeMarkdownStylingKt {
 	public static final fun create (Lorg/jetbrains/jewel/markdown/rendering/InlinesStyling$Companion;Landroidx/compose/ui/text/TextStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Z)Lorg/jetbrains/jewel/markdown/rendering/InlinesStyling;
 	public static final fun create (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code$Companion;Landroidx/compose/ui/text/TextStyle;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code$Indented;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code$Fenced;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code;
 	public static final fun create (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$Companion;Landroidx/compose/ui/text/TextStyle;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H1;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H2;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H3;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H4;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H5;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H6;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading;
@@ -48,7 +51,7 @@ public final class org/jetbrains/jewel/intui/markdown/styling/MarkdownBridgeStyl
 	public static final fun retrieveEditorColorScheme ()Lcom/intellij/openapi/editor/colors/EditorColorsScheme;
 }
 
-public final class org/jetbrains/jewel/intui/markdown/styling/extension/github/alerts/GitHubAlertBridgeStylingKt {
+public final class org/jetbrains/jewel/intui/markdown/bridge/styling/extension/github/alerts/BridgeGitHubAlertStylingKt {
 	public static final fun create (Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling$Companion;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/NoteAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/TipAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/ImportantAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/WarningAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling;)Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling;
 	public static synthetic fun create$default (Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling$Companion;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/NoteAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/TipAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/ImportantAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/WarningAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling;ILjava/lang/Object;)Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling;
 	public static final fun create-gaOEZmc (Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling$Companion;Landroidx/compose/foundation/layout/PaddingValues;FJLandroidx/compose/ui/graphics/PathEffect;ILandroidx/compose/ui/text/TextStyle;Ljava/lang/String;JJ)Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling;
