+++ markdown/ide-laf-bridge-styling/src/main/kotlin/org/jetbrains/jewel/intui/markdown/bridge/BridgeMarkdownBlockRendererExtensions.kt
@@ -1,0 +1,14 @@
+package org.jetbrains.jewel.intui.markdown.bridge
+
+import org.jetbrains.jewel.intui.markdown.bridge.styling.create
+import org.jetbrains.jewel.markdown.extensions.MarkdownRendererExtension
+import org.jetbrains.jewel.markdown.rendering.DefaultMarkdownBlockRenderer
+import org.jetbrains.jewel.markdown.rendering.InlineMarkdownRenderer
+import org.jetbrains.jewel.markdown.rendering.MarkdownBlockRenderer
+import org.jetbrains.jewel.markdown.rendering.MarkdownStyling
+
+public fun MarkdownBlockRenderer.Companion.create(
+    styling: MarkdownStyling = MarkdownStyling.create(),
+    rendererExtensions: List<MarkdownRendererExtension> = emptyList(),
+    inlineRenderer: InlineMarkdownRenderer = InlineMarkdownRenderer.default(),
+): MarkdownBlockRenderer = DefaultMarkdownBlockRenderer(styling, rendererExtensions, inlineRenderer)
