+++ markdown/ide-laf-bridge-styling/src/main/kotlin/org/jetbrains/jewel/intui/markdown/bridge/styling/extension/github/alerts/BridgeGitHubAlertStyling.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.jewel.intui.markdown.styling.extension.github.alerts
+package org.jetbrains.jewel.intui.markdown.bridge.styling.extension.github.alerts
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.ui.graphics.Color
