+++ markdown/int-ui-standalone-styling/api/int-ui-standalone-styling.api
@@ -1,13 +1,16 @@
-public final class org/jetbrains/jewel/intui/markdown/IntUiMarkdownBlockRendererExtensionsKt {
-	public static final fun ProvideMarkdownStyling (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;Lorg/jetbrains/jewel/markdown/processing/MarkdownProcessor;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
-	public static final fun ProvideMarkdownStyling (ZLorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Lorg/jetbrains/jewel/markdown/processing/MarkdownProcessor;Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
+public final class org/jetbrains/jewel/intui/markdown/standalone/IntUiMarkdownBlockRendererExtensionsKt {
 	public static final fun dark (Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Ljava/util/List;Lorg/jetbrains/jewel/markdown/rendering/InlineMarkdownRenderer;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;
 	public static synthetic fun dark$default (Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Ljava/util/List;Lorg/jetbrains/jewel/markdown/rendering/InlineMarkdownRenderer;ILjava/lang/Object;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;
 	public static final fun light (Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Ljava/util/List;Lorg/jetbrains/jewel/markdown/rendering/InlineMarkdownRenderer;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;
 	public static synthetic fun light$default (Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Ljava/util/List;Lorg/jetbrains/jewel/markdown/rendering/InlineMarkdownRenderer;ILjava/lang/Object;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;
 }
 
-public final class org/jetbrains/jewel/intui/markdown/styling/MarkdownIntUiStylingKt {
+public final class org/jetbrains/jewel/intui/markdown/standalone/IntUiProvideMarkdownStylingKt {
+	public static final fun ProvideMarkdownStyling (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;Lorg/jetbrains/jewel/markdown/processing/MarkdownProcessor;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
+	public static final fun ProvideMarkdownStyling (ZLorg/jetbrains/jewel/markdown/rendering/MarkdownStyling;Lorg/jetbrains/jewel/markdown/processing/MarkdownProcessor;Lorg/jetbrains/jewel/markdown/rendering/MarkdownBlockRenderer;Lkotlin/jvm/functions/Function2;Landroidx/compose/runtime/Composer;II)V
+}
+
+public final class org/jetbrains/jewel/intui/markdown/standalone/styling/IntUiMarkdownStylingKt {
 	public static final fun dark (Lorg/jetbrains/jewel/markdown/rendering/InlinesStyling$Companion;Landroidx/compose/ui/text/TextStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Landroidx/compose/ui/text/SpanStyle;Z)Lorg/jetbrains/jewel/markdown/rendering/InlinesStyling;
 	public static final fun dark (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code$Indented;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code$Fenced;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Code;
 	public static final fun dark (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$Companion;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H1;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H2;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H3;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H4;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H5;Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading$H6;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$Heading;
@@ -88,7 +91,7 @@ public final class org/jetbrains/jewel/intui/markdown/styling/MarkdownIntUiStyli
 	public static synthetic fun light-pI2OzKA$default (Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$BlockQuote$Companion;Landroidx/compose/foundation/layout/PaddingValues;FJLandroidx/compose/ui/graphics/PathEffect;IJILjava/lang/Object;)Lorg/jetbrains/jewel/markdown/rendering/MarkdownStyling$BlockQuote;
 }
 
-public final class org/jetbrains/jewel/intui/markdown/styling/extension/github/alerts/GitHubAlertDefaultStylingKt {
+public final class org/jetbrains/jewel/intui/markdown/standalone/styling/extension/github/alerts/IntUiGitHubAlertStylingKt {
 	public static final fun dark (Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling$Companion;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/NoteAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/TipAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/ImportantAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/WarningAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling;)Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling;
 	public static synthetic fun dark$default (Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling$Companion;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/NoteAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/TipAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/ImportantAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/WarningAlertStyling;Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling;ILjava/lang/Object;)Lorg/jetbrains/jewel/markdown/extensions/github/alerts/AlertStyling;
 	public static final fun dark-gaOEZmc (Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling$Companion;Landroidx/compose/foundation/layout/PaddingValues;FJLandroidx/compose/ui/graphics/PathEffect;ILandroidx/compose/ui/text/TextStyle;Ljava/lang/String;JJ)Lorg/jetbrains/jewel/markdown/extensions/github/alerts/CautionAlertStyling;
