+++ samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/view/markdown/MarkdownPreview.kt
@@ -20,13 +20,13 @@ import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.jetbrains.jewel.foundation.theme.JewelTheme
-import org.jetbrains.jewel.intui.markdown.ProvideMarkdownStyling
-import org.jetbrains.jewel.intui.markdown.dark
-import org.jetbrains.jewel.intui.markdown.light
-import org.jetbrains.jewel.intui.markdown.styling.dark
-import org.jetbrains.jewel.intui.markdown.styling.extension.github.alerts.dark
-import org.jetbrains.jewel.intui.markdown.styling.extension.github.alerts.light
-import org.jetbrains.jewel.intui.markdown.styling.light
+import org.jetbrains.jewel.intui.markdown.standalone.ProvideMarkdownStyling
+import org.jetbrains.jewel.intui.markdown.standalone.dark
+import org.jetbrains.jewel.intui.markdown.standalone.light
+import org.jetbrains.jewel.intui.markdown.standalone.styling.dark
+import org.jetbrains.jewel.intui.markdown.standalone.styling.extension.github.alerts.dark
+import org.jetbrains.jewel.intui.markdown.standalone.styling.extension.github.alerts.light
+import org.jetbrains.jewel.intui.markdown.standalone.styling.light
 import org.jetbrains.jewel.markdown.MarkdownBlock
 import org.jetbrains.jewel.markdown.extension.autolink.AutolinkProcessorExtension
 import org.jetbrains.jewel.markdown.extensions.LazyMarkdown
